refactor(lint-cli): group the implementation under lib/ by concern
A flat 28-file directory said nothing about what belonged with what. The two
program entry points (index.ts, ignored-child.ts, both tsdown entries) stay at
the root; everything else moves under lib/ grouped by concern: cli, files,
cache, typescript, hybrid, plan, exec.

Three files that spanned concerns are split along them:
- constants.ts was a cross-concern dumping ground. Cache file names and the
  bust threshold go to cache/, the repo-scan globs and extensions to files/,
  and the worker tuning constants into plan/concurrency.ts beside the only
  code that reads them.
- command.ts held argument splitting, command composition and shell quoting.
  They become cli/split-args.ts, plan/command.ts and exec/shell.ts.
- hybrid.ts held both the run decision and the child spawn it depends on;
  probe.ts is now the injected seam gate.ts already treated it as.

resolveIgnoredHelper's source fallback is two directories deeper and updated
to match. That path fails silently by design (a broken one degrades to "no
ignore filtering"), so it was verified by breaking it deliberately: three
lint-cli-ignored tests fail, confirming the suite covers it.
