+++ src/lint-cli/lib/cache/file-entry-cache.d.ts
