+++ src/lint-cli/lib/exec/resolve.ts
@@ -3,7 +3,7 @@ import fs from "node:fs";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 
-import { CliError } from "./types.ts";
+import { CliError } from "../cli/types.ts";
 
 /** Memoised {@link resolveLocalBin} results, keyed by `cwd\0name`. */
 const localBinCache = new Map<string, string>();
@@ -49,6 +49,13 @@ export function resolveLocalBin(name: string, cwd: string): string {
  * (built from `src/formatter-agents.ts` to `formatter-agents.mjs`). Resolved
  * lazily so it is only touched when `--agents` is used.
  *
+ * Meaningful only in the bundled layout, where every entry is flattened into
+ * `dist/` and this resolves to `dist/formatter-agents.mjs`. Run from source
+ * there is no sibling `.mjs` at all, so the path names a file that does not
+ * exist — `--agents` is a shipped-package feature and has never worked from a
+ * source checkout. Unlike {@link resolveIgnoredHelper} there is no source
+ * fallback, because ESLint would have to load the `.ts` formatter itself.
+ *
  * @returns The absolute path to the agent ESLint formatter.
  */
 export function resolveAgentsFormatter(): string {
@@ -60,10 +67,13 @@ export function resolveAgentsFormatter(): string {
  * `src/lint-cli/ignored-child.ts` to `lint-ignored.mjs`).
  *
  * Unlike {@link resolveAgentsFormatter} this falls back to the TypeScript
- * source sibling, because the runner spawns it itself and the fixture tests run
- * from `src` where no bundle exists. The fallback means the tests never
- * exercise the shipped path — a broken or unshipped `dist` entry degrades the
- * runner to "no ignore filtering" rather than failing.
+ * source, because the runner spawns it itself and the fixture tests run from
+ * `src` where no bundle exists. That fallback path is relative to *this
+ * file's* source location (`lib/exec/`), while the built branch resolves
+ * against a flat `dist/` — which is why only the former carries `../../`. The
+ * fallback means the tests never exercise the shipped path — a broken or
+ * unshipped `dist` entry degrades the runner to "no ignore filtering" rather
+ * than failing.
  *
  * @returns The absolute path to the helper module.
  */
@@ -71,5 +81,5 @@ export function resolveIgnoredHelper(): string {
 	const built = fileURLToPath(new URL("./lint-ignored.mjs", import.meta.url));
 	return fs.existsSync(built)
 		? built
-		: fileURLToPath(new URL("./ignored-child.ts", import.meta.url));
+		: fileURLToPath(new URL("../../ignored-child.ts", import.meta.url));
 }
