+++ src/lint-cli/lib/exec/shell.ts
@@ -1,138 +1,10 @@
-import type {
-	ChildCommand,
-	ComposeContext,
-	LintCliOptions,
-	OxlintComposeContext,
-} from "./types.ts";
-import { CliError } from "./types.ts";
+import type { ChildCommand } from "../cli/types.ts";
+import { CliError } from "../cli/types.ts";
 
 // `%` is deliberately excluded: cmd.exe expands `%VAR%` even inside double
 // quotes, so a `%` token cannot be made safe by quoting on the Windows shell
 // path (see `buildShellCommand`).
 const SAFE_TOKEN = /^[\w@+=:,./-]+$/;
-
-/**
- * Split a raw argument string into tokens, honouring single and double quotes
- * so values such as `--rule "no-console: error"` survive as one argument.
- *
- * @param input - The raw argument string to split.
- * @returns The parsed argument tokens.
- */
-export function splitArgs(input: string): Array<string> {
-	const tokens: Array<string> = [];
-	let current = "";
-	let quote: "'" | '"' | undefined;
-	let hasToken = false;
-
-	for (const char of input) {
-		if (quote !== undefined) {
-			if (char === quote) {
-				quote = undefined;
-			} else {
-				current += char;
-			}
-
-			continue;
-		}
-
-		if (char === '"' || char === "'") {
-			quote = char;
-			hasToken = true;
-			continue;
-		}
-
-		if (char === " " || char === "\t" || char === "\n") {
-			if (hasToken) {
-				tokens.push(current);
-				current = "";
-				hasToken = false;
-			}
-
-			continue;
-		}
-
-		current += char;
-		hasToken = true;
-	}
-
-	if (hasToken) {
-		tokens.push(current);
-	}
-
-	return tokens;
-}
-
-/**
- * Compose the oxlint child command.
- *
- * @param options - The parsed CLI options.
- * @param context - The resolved composition context.
- * @returns The oxlint child command.
- */
-export function composeOxlintCommand(
-	options: LintCliOptions,
-	context: OxlintComposeContext,
-): ChildCommand {
-	const args: Array<string> = [];
-	if (options.agents) {
-		args.push("--format", "agent");
-	}
-
-	if (context.oxlintTypeAware) {
-		args.push("--type-aware");
-	}
-
-	if (options.fix) {
-		args.push("--fix");
-	}
-
-	args.push(...options.oxlintArgs, ...context.paths);
-	return { args, bin: "oxlint", env: {}, label: "oxc" };
-}
-
-/**
- * Compose the ESLint child command.
- *
- * @param options - The parsed CLI options.
- * @param context - The resolved composition context.
- * @returns The ESLint child command.
- */
-export function composeEslintCommand(
-	options: LintCliOptions,
-	context: ComposeContext,
-): ChildCommand {
-	const args: Array<string> = [];
-	if (options.cache) {
-		args.push("--cache", "--cache-location", context.cacheLocation);
-	}
-
-	args.push("--no-warn-ignored", "--concurrency", String(context.concurrency));
-
-	if (options.cache && context.ci) {
-		args.push("--cache-strategy", "content");
-	}
-
-	if (options.agents) {
-		args.push("--format", context.agentsFormatterPath);
-	}
-
-	if (options.fix) {
-		args.push("--fix");
-	}
-
-	args.push(...options.eslintArgs, ...context.paths);
-
-	// Always control ESLINT_TYPE_AWARE explicitly: set it for the fast/typed
-	// passes, and set it to `undefined` for the full pass so an inherited value
-	// (a user-exported ESLINT_TYPE_AWARE) is REMOVED rather than leaked — a leak
-	// would make the factory split the full config (e.g. --fix applying only
-	// type-aware fixes).
-	const environment: Record<string, string | undefined> = {
-		ESLINT_TYPE_AWARE: context.typeAwareEnv,
-	};
-
-	return { args, bin: "eslint", env: environment, label: context.eslintLabel };
-}
 
 /**
  * Render a command as a shell-equivalent line (env prefix + logical binary +
