+++ src/lint-cli/lib/paths.ts
