+++ src/lint-cli/lib/state.ts
