+++ src/lint-cli/lib/typescript/load.ts
